REF: Move `ServicePatcher` into its own module
