+++ tests/base.py
@@ -6,8 +6,8 @@ import yaml
 
 from octue.cloud import storage
 from octue.cloud.emulators._pub_sub import MockTopic
-from octue.cloud.emulators.child import ServicePatcher
 from octue.cloud.emulators.cloud_storage import GoogleCloudStorageEmulatorTestResultModifier
+from octue.cloud.emulators.service import ServicePatcher
 from octue.cloud.events import OCTUE_SERVICES_PREFIX
 from octue.cloud.storage import GoogleCloudStorageClient
 from octue.resources import Datafile, Dataset, Manifest
